feat(time): move duration implementation and tests into dedicated files
